# TM8/tm8-game-mobile-release-0.6.0/input_folder/dev/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/statistics/users/total-count": {
            "get": {
                "operationId": "StatisticsController_getUserCount",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatisticsTotalCountResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Statistics"
                ]
            }
        },
        "/api/v1/statistics/users/onboarding-completion": {
            "get": {
                "operationId": "StatisticsController_getOnboardingCompletion",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatisticsOnboardingCompletionResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Statistics"
                ]
            }
        },
        "/api/v1/statistics/users/new-users-registered": {
            "get": {
                "operationId": "StatisticsController_getNewUsersRegistered",
                "parameters": [
                    {
                        "name": "groupBy",
                        "required": true,
                        "in": "query",
                        "description": "Group parameter",
                        "schema": {
                            "enum": [
                                "day",
                                "week",
                                "month",
                                "year"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "startDate",
                        "required": true,
                        "in": "query",
                        "description": "Start date",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endDate",
                        "required": true,
                        "in": "query",
                        "description": "End date",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatisticsNewUsersRegisteredResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Statistics"
                ]
            }
        },
        "/api/v1/statistics/users/group-counts": {
            "get": {
                "operationId": "StatisticsController_getUserGroupsCounts",
                "parameters": [
                    {
                        "name": "userGroups",
                        "required": false,
                        "in": "query",
                        "description": "User group filter",
                        "schema": {
                            "nullable": true,
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGroupCountsResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Statistics"
                ]
            }
        },
...
    },
    "info": {
        "title": "TM8 API",
        "description": "TM8 API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "TGMDExceptionResponse": {
                "type": "object",
                "properties": {
                    "origin": {
                        "enum": [
                            "HTTP",
                            "Mongo",
                            "TGMD",
                            "Unknown"
                        ],
                        "type": "string"
                    },
                    "status": {
                        "type": "number",
                        "description": "HTTP-compatible status code"
                    },
                    "code": {
                        "type": "string"
                    },
                    "exception": {
                        "type": "string",
                        "description": "Exception name"
                    },
                    "detail": {
                        "type": "string",
                        "description": "Human-readable error message"
                    }
                },
                "required": [
                    "origin",
                    "status",
                    "code",
                    "exception",
                    "detail"
                ]
            },
...
            "StatisticsTotalCountResponse": {
                "type": "object",
                "properties": {
                    "total": {
                        "type": "number"
                    },
                    "currentWeek": {
                        "type": "number"
                    }
                },
                "required": [
                    "total",
                    "currentWeek"
                ]
            },
            "StatisticsOnboardingCompletionResponse": {
                "type": "object",
                "properties": {
                    "onboardingPct": {
                        "type": "number"
                    },
                    "currentWeek": {
                        "type": "number"
                    }
                },
                "required": [
                    "onboardingPct",
                    "currentWeek"
                ]
            },
            "NewUsersRegisteredObjectResponse": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number"
                    }
                },
                "required": [
                    "date",
                    "quantity"
                ]
            },
            "StatisticsNewUsersRegisteredResponse": {
                "type": "object",
                "properties": {
                    "chart": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NewUsersRegisteredObjectResponse"
                        }
                    }
                },
                "required": [
                    "chart"
                ]
            },
            "UserGroupCountsResponse": {
                "type": "object",
                "properties": {
                    "allUsers": {
                        "type": "number",
                        "nullable": true
                    },
                    "apexLegendsPlayers": {
                        "type": "number",
                        "nullable": true
                    },
                    "callOfDutyPlayers": {
                        "type": "number",
                        "nullable": true
                    },
                    "fortnitePlayers": {
                        "type": "number",
                        "nullable": true
                    },
                    "rocketLeaguePlayers": {
                        "type": "number",
                        "nullable": true
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}